Change fetch building logic. Update page rerender on building addition
